Fixed whatever was wrong with bullets.
It was a stupid thing, but bleh.

Also, shooting any of the obstacles in the main scene changes the color
of said object randomly (player feedback to test collision)

slightly adjusted the player's cannons to be more accurate in longer
ranges (it's static for now, but I'm gonna make it rotate according to
the reticle).